requirements.txt
Pillow==11.2.1
tk==0.1.0
